Add files via upload
League of Legends RP Conversion Calculator

This simple program allows you to convert a specified amount of Euros (EUR) into Riot Points (RP) for League of Legends. By inputting the amount of Euros you wish to convert, the program will use the current exchange rate to calculate and display the equivalent amount of RP you can buy.

The exchange rate used is approximately 1 RP = 0.0063 EUR, but please note that the actual rate may vary depending on the region and time.