Add files via upload
pH Level Checker (Learning  If/Elif/Else)

This simple program helps to determine whether a liquid is acidic, basic, or neutral based on its pH level. By inputting a pH value between 0 and 14, the program will classify it as:

Basic: If the pH is greater than 7.
Acidic: If the pH is less than 7.
Neutral: If the pH equals 7.