3.Project Venus.py
Age Calculator with Venus Years
This Python program allows the user to input their name and birth year, then calculates their age based on the current year (2025).

But wait, there's more! You can also find out how old you'd be if you lived on Venus! 🌍🪐 A year on Venus is only 0.62 times the length of a year on Earth, so this program will convert your age to Venus years as well.

The program includes the following:

Asking for the user's name.
Asking for the user's birth year and calculating their age.
Giving the user the option to calculate their age in Venus years.
If the user decides they don't want to know their age in Venus years, the program will respond with a fun pun to keep things lighthearted!